+++ restio/transaction.py
@@ -404,6 +404,15 @@ class Transaction:
                 )
 
     def _update(self, model: BaseModel, field: Field[T_co], value: T_co) -> bool:
+        # check if any value being passed is actually a model that is already
+        # registered in the cache, otherwise complains
+        if field.depends_on:
+            if isinstance(value, Iterable):
+                for v in value:
+                    self._check_model_is_valid(v)
+            elif isinstance(value, BaseModel):
+                self._check_model_is_valid(value)
+
         # During ROLLBACK and GET, all changes to models should be permanent, therefore
         # we persist any change and avoid changing the state of the models
         if self.state in (TransactionState.ROLLBACK, TransactionState.GET):
@@ -495,16 +504,7 @@ class Transaction:
                 " registered again to a Transaction."
             )
 
-        children = model.get_children(recursive=True)
-        for child in children:
-            if not self._model_cache.get_by_internal_id(
-                child.__class__, child._internal_id
-            ):
-                raise RuntimeError(
-                    f"Model of id `{child._internal_id}` ({child.__class__.__name__}) "
-                    f"is a child of `{model._internal_id}` ({model.__class__.__name__})"
-                    " and is not registered in cache."
-                )
+        self._check_models_children({model}, recursive=True)
 
         registered = self._model_cache.register(model, force)
         if registered:
@@ -582,21 +582,20 @@ class Transaction:
         :return: A list containing all DAOTask's performed by the commit, in the order
                  in which the operations have been finalized.
         """
-        cached_values = self._model_cache.get_all_models()
+        cached_models = self._model_cache.get_all_models()
 
-        models_to_add = self._get_models_to_add(cached_values)
-        models_to_update = self._get_models_to_update(cached_values)
-        models_to_remove = self._get_models_to_remove(cached_values)
+        self._check_models_consistency(cached_models)
+
+        models_to_add = self._get_models_to_add(cached_models)
+        models_to_update = self._get_models_to_update(cached_models)
+        models_to_remove = self._get_models_to_remove(cached_models)
 
         all_models: Set[BaseModel] = models_to_add.union(models_to_update).union(
             models_to_remove
         )
 
-        self._check_deleted_models(cached_values)
-
-        # add, update and remove get one graph each, as
-        # each level needs to be completely finished in
-        # order for the next to proceed
+        # add, update and remove get one graph each, as each level needs to be
+        # completely finished in order for the next to proceed
         graphs = [
             DependencyGraph.generate_from_objects(m)
             for m in (models_to_add, models_to_update, models_to_remove)
@@ -622,10 +621,43 @@ class Transaction:
             self.update_cache()
         return results
 
+    def _check_models_consistency(self, models: Set[ModelType]):
+        for model in models:
+            self._check_model_is_valid(model)
+
+        self._check_models_children(models, recursive=False)
+
+        self._check_deleted_models(models)
+
+    def _check_models_children(self, models: Set[ModelType], recursive: bool):
+        for model in models:
+            for child in model.get_children(recursive=recursive):
+                self._check_model_is_valid(child)
+
+    def _check_model_is_valid(self, model: ModelType):
+        status_valid = model._state not in (ModelState.DISCARDED, ModelState.UNBOUND,)
+
+        if not status_valid:
+            raise ValueError(
+                f"Model of id `{model._internal_id}` ({model.__class__.__name__}) is"
+                f" not valid because its state is either DISCARDED or UNBOUND."
+            )
+
+        found_in_cache = bool(
+            self._model_cache.get_by_internal_id(model.__class__, model._internal_id)
+        )
+
+        if not found_in_cache:
+            raise ValueError(
+                f"Model of id `{model._internal_id}` ({model.__class__.__name__}) is"
+                f" not valid because it is not registered to the cache."
+            )
+
     @staticmethod
     def _check_deleted_models(models: Set[BaseModel]):
         nodes = DependencyGraph._get_connected_nodes(models)
-        for node in (n for n in nodes if n.node_object._state == ModelState.DELETED):
+        deleted_nodes = (n for n in nodes if n.node_object._state == ModelState.DELETED)
+        for node in deleted_nodes:
             for parent_node in node.get_parents(recursive=True):
                 parent_model: BaseModel = parent_node.node_object
                 if parent_model._state not in (